skip processing if no asset data was found
For a team we should always get asset data back from ES. In the case of
all issues resolved, this would just be a set of assets with an empty
vulnerabilities list in each one. If we get nothing, just skip
processing as it's possible something went wrong when information was
being updated in the index.